refactor: `setSingleClickListener` 확장 함수 위치 변경
