+++ android/app/src/main/java/com/mulkkam/ui/custom/ExtendableFloatingActionButton.kt
@@ -14,7 +14,7 @@ import androidx.core.view.children
 import androidx.core.view.isVisible
 import com.mulkkam.R
 import com.mulkkam.databinding.LayoutExpandableFloatingActionButtonBinding
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class ExtendableFloatingActionButton
     @JvmOverloads
