+++ android/app/src/main/java/com/mulkkam/ui/home/dialog/ManualDrinkFragment.kt
@@ -8,7 +8,7 @@ import com.mulkkam.domain.model.IntakeType
 import com.mulkkam.ui.binding.BindingBottomSheetDialogFragment
 import com.mulkkam.ui.custom.MulKkamChipGroupAdapter
 import com.mulkkam.ui.home.HomeViewModel
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class ManualDrinkFragment :
     BindingBottomSheetDialogFragment<FragmentManualDrinkBinding>(
