+++ android/app/src/main/java/com/mulkkam/ui/notification/NotificationActivity.kt
@@ -9,7 +9,7 @@ import com.mulkkam.R
 import com.mulkkam.databinding.ActivityNotificationBinding
 import com.mulkkam.ui.binding.BindingActivity
 import com.mulkkam.ui.notification.adapter.NotificationAdapter
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class NotificationActivity :
     BindingActivity<ActivityNotificationBinding>(
