+++ android/app/src/main/java/com/mulkkam/ui/notification/adapter/NotificationViewHolder.kt
@@ -10,7 +10,7 @@ import com.mulkkam.R
 import com.mulkkam.databinding.ItemHomeNotificationBinding
 import com.mulkkam.domain.model.Notification
 import com.mulkkam.domain.model.NotificationType.SUGGESTION
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 import java.time.Duration
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
