+++ android/app/src/main/java/com/mulkkam/ui/onboarding/bioinfo/BioInfoFragment.kt
@@ -15,8 +15,7 @@ import com.mulkkam.ui.binding.BindingFragment
 import com.mulkkam.ui.onboarding.OnboardingViewModel
 import com.mulkkam.ui.onboarding.bioinfo.dialog.OnboardingWeightFragment
 import com.mulkkam.ui.util.extensions.getAppearanceSpannable
-import com.mulkkam.ui.util.getAppearanceSpannable
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class BioInfoFragment :
     BindingFragment<FragmentBioInfoBinding>(
