+++ android/app/src/main/java/com/mulkkam/ui/onboarding/bioinfo/dialog/OnboardingWeightFragment.kt
@@ -9,7 +9,7 @@ import com.mulkkam.domain.model.BioWeight.Companion.WEIGHT_MAX
 import com.mulkkam.domain.model.BioWeight.Companion.WEIGHT_MIN
 import com.mulkkam.ui.binding.BindingBottomSheetDialogFragment
 import com.mulkkam.ui.onboarding.bioinfo.BioInfoViewModel
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class OnboardingWeightFragment :
     BindingBottomSheetDialogFragment<FragmentWeightBinding>(
