+++ android/app/src/main/java/com/mulkkam/ui/onboarding/nickname/NicknameFragment.kt
@@ -15,7 +15,7 @@ import com.mulkkam.ui.binding.BindingFragment
 import com.mulkkam.ui.onboarding.OnboardingViewModel
 import com.mulkkam.ui.util.extensions.applyImeMargin
 import com.mulkkam.ui.util.extensions.getAppearanceSpannable
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class NicknameFragment :
     BindingFragment<FragmentNicknameBinding>(
