+++ android/app/src/main/java/com/mulkkam/ui/onboarding/targetamount/TargetAmountFragment.kt
@@ -17,7 +17,7 @@ import com.mulkkam.ui.onboarding.OnboardingViewModel
 import com.mulkkam.ui.util.extensions.applyImeMargin
 import com.mulkkam.ui.util.extensions.getAppearanceSpannable
 import com.mulkkam.ui.util.extensions.getColoredSpannable
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 import java.util.Locale
 
 class TargetAmountFragment :
