+++ android/app/src/main/java/com/mulkkam/ui/onboarding/terms/TermsFragment.kt
@@ -13,7 +13,7 @@ import com.mulkkam.ui.onboarding.OnboardingViewModel
 import com.mulkkam.ui.onboarding.terms.adapter.TermsAdapter
 import com.mulkkam.ui.util.extensions.applyImeMargin
 import com.mulkkam.ui.util.extensions.getAppearanceSpannable
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 import kotlin.collections.find
 
 class TermsFragment :
