+++ android/app/src/main/java/com/mulkkam/ui/settingcups/adapter/CupViewHolder.kt
@@ -9,7 +9,7 @@ import com.mulkkam.databinding.ItemSettingCupsCupBinding
 import com.mulkkam.domain.model.IntakeType
 import com.mulkkam.ui.settingcups.model.CupUiModel
 import com.mulkkam.ui.util.extensions.loadUrl
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 
 class CupViewHolder(
     parent: ViewGroup,
