+++ android/app/src/main/java/com/mulkkam/ui/settingcups/dialog/SettingCupFragment.kt
@@ -14,7 +14,7 @@ import com.mulkkam.ui.settingcups.SettingCupsViewModel
 import com.mulkkam.ui.settingcups.model.CupUiModel
 import com.mulkkam.ui.settingcups.model.CupUiModel.Companion.EMPTY_CUP_UI_MODEL
 import com.mulkkam.ui.settingcups.model.SettingWaterCupEditType
-import com.mulkkam.ui.util.setSingleClickListener
+import com.mulkkam.ui.util.extensions.setSingleClickListener
 import com.mulkkam.util.extensions.getParcelableCompat
 
 class SettingCupFragment :
